refactore code created FeedbackResponseController
