+++ app/views/header.scala.html
@@ -49,7 +49,7 @@
                     } else {
 
                         <li class="@activeLink(routes.UsersController.register().url) alignLinks">
-                            <a href="@routes.FeedbackFormsController.getFeedbackFormsForToday"><i class="fa fa-comments"></i> Feedbacks</a>
+                            <a href="@routes.FeedbackFormsResponseController.getFeedbackFormsForToday"><i class="fa fa-comments"></i> Feedbacks</a>
                         </li>
 
                         <li class="dropdown alignLinks">
