+++ conf/routes
@@ -31,7 +31,7 @@ GET         /feedbackform/update                 controllers.FeedbackFormsContro
 POST        /feedbackform/update                 controllers.FeedbackFormsController.updateFeedbackForm
 GET         /feedbackform/delete                 controllers.FeedbackFormsController.deleteFeedbackForm(id: String)
 GET         /feedbackform/preview/:id            controllers.FeedbackFormsController.getFeedbackFormPreview(id: String)
-GET         /feedbackform/for/today              controllers.FeedbackFormsController.getFeedbackFormsForToday
+GET         /feedbackform/for/today              controllers.FeedbackFormsResponseController.getFeedbackFormsForToday
 GET         /jsRoutes                            controllers.JavascriptRouter.jsRoutes
 # Map static resources from the /public folder to the /assets URL path
 GET         /assets/*file                        controllers.Assets.versioned(path="/public", file: Asset)
