+++ test/controllers/FeedbackFormsControllerSpec.scala
@@ -1,7 +1,6 @@
 package controllers
 
 import java.text.SimpleDateFormat
-import java.time.ZoneId
 
 import com.typesafe.config.ConfigFactory
 import models._
@@ -398,44 +397,6 @@ class FeedbackFormsControllerSpec extends PlaySpecification with TestEnvironment
       mailerClient.send(email) returns ""
 
       val response = controller.sendFeedbackForm(_id.stringify)(FakeRequest()
-        .withSession("username" -> "uNtgSXeM+2V+h8ChQT/PiHq70PfDk+sGdsYAXln9GfU="))
-
-      status(response) must be equalTo OK
-    }
-    "not render feedback form for today if session associated feedback form not found" in new WithTestApplication {
-      usersRepository.getByEmail("test@example.com") returns emailObject
-      sessionsRepository.getSessionsTillNow returns sessionObject
-      feedbackFormsRepository.getByFeedbackFormId("feedbackFormId") returns Future.successful(None)
-      dateTimeUtility.ISTZoneId returns ZoneId.of("Asia/Calcutta")
-      dateTimeUtility.nowMillis returns date.getTime
-      val response = controller.getFeedbackFormsForToday(FakeRequest()
-        .withSession("username" -> "uNtgSXeM+2V+h8ChQT/PiHq70PfDk+sGdsYAXln9GfU="))
-
-      status(response) must be equalTo OK
-    }
-
-    "render feedback form for today if session associated feedback form found and session not expired" in new WithTestApplication {
-      val sessionObjectWithCurrentDate =
-        Future.successful(List(SessionInfo(_id.stringify, "email", BSONDateTime(System.currentTimeMillis), "sessions", "feedbackFormId", "topic",
-          1, meetup = true, "rating", cancelled = false, active = true, _id)))
-      usersRepository.getByEmail("test@example.com") returns emailObject
-      sessionsRepository.getSessionsTillNow returns sessionObjectWithCurrentDate
-      feedbackFormsRepository.getByFeedbackFormId("feedbackFormId") returns Future.successful(Some(feedbackForms))
-      dateTimeUtility.ISTZoneId returns ZoneId.of("Asia/Calcutta")
-      dateTimeUtility.nowMillis returns date.getTime
-      val response = controller.getFeedbackFormsForToday(FakeRequest()
-        .withSession("username" -> "uNtgSXeM+2V+h8ChQT/PiHq70PfDk+sGdsYAXln9GfU="))
-
-      status(response) must be equalTo OK
-    }
-
-    "render feedback form for today if session associated feedback form found and session expired" in new WithTestApplication {
-      usersRepository.getByEmail("test@example.com") returns emailObject
-      sessionsRepository.getSessionsTillNow returns sessionObject
-      feedbackFormsRepository.getByFeedbackFormId("feedbackFormId") returns Future.successful(Some(feedbackForms))
-      dateTimeUtility.ISTZoneId returns ZoneId.of("Asia/Calcutta")
-      dateTimeUtility.nowMillis returns date.getTime
-      val response = controller.getFeedbackFormsForToday(FakeRequest()
         .withSession("username" -> "uNtgSXeM+2V+h8ChQT/PiHq70PfDk+sGdsYAXln9GfU="))
 
       status(response) must be equalTo OK
